requirements.txt
appdirs==1.4.4
attrs==23.1.0
backoff==2.2.1
boilerpy3==1.0.6
canals==0.7.0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
docopt==0.6.2
Events==0.5
exceptiongroup==1.1.3
faiss-cpu==1.7.2
farm-haystack==1.20.1
filelock==3.12.3
fsspec==2023.9.0
greenlet==2.0.2
huggingface-hub==0.17.1
idna==3.4
inflect==7.0.0
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy-imports==0.3.1
monotonic==1.6
more-itertools==10.1.0
networkx==3.1
num2words==0.5.12
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
platformdirs==3.10.0
posthog==3.0.2
prompthub-py==4.0.0
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.10.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.41.1
sseclient-py==1.8.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.7.1
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.4
